улучшение кода
