+++ js/upload_form.js
@@ -1,26 +1,6 @@
-import { displayElementRemove, displayElementAdd } from './big_picture.js';
-import { isEscapeKey } from './util.js';
-import { cleaningEffects } from './effects_photo.js';
+import { displayElementAdd, displayElementRemove } from './util.js';
 import { rescalingPhoto, defaultSettingsFilter } from './photo_resizing.js';
-import { defaultSettingsSlider } from './slider.js';
-
-let removeEditWindow = null;
-
-let removeEditWindow2 = null;
-
-const closeEditWindow = () => {
-  document.addEventListener('keydown', removeEditWindow);
-};
-
-const closeEditWindow2 = () => {
-  document.querySelector('#upload-cancel').addEventListener('click', removeEditWindow2);
-};
-
-const closeEditForm = () => {
-  closeEditWindow();
-  closeEditWindow2();
-  defaultSettingsSlider();
-};
+import {closeEditForm} from './close_form.js';
 
 const openingUploadForm = () => {
   displayElementRemove('.img-upload__overlay', 'hidden');
@@ -28,24 +8,6 @@ const openingUploadForm = () => {
   closeEditForm();
   defaultSettingsFilter(100);
   rescalingPhoto(100);
-};
-
-const removeEventListener = () => {
-  document.addEventListener('keydown', removeEditWindow);
-  document.querySelector('#upload-cancel').addEventListener('click', removeEditWindow2);
-};
-
-removeEditWindow2 = () => {
-  displayElementAdd('.img-upload__overlay', 'hidden');
-  displayElementRemove('body', 'modal-open');
-  cleaningEffects();
-  removeEventListener();
-};
-
-removeEditWindow = function (evt) {
-  if (isEscapeKey(evt)) {
-    removeEditWindow2();
-  }
 };
 
 export const uploadNewImage = () => {
